Compressed questions into Object
Switched from having a whole function and then some for a questions, to
adding the variables to a object within an other object so that any
question's data can be accessed at any time.

This also means that it reduces the amount of work to add a question.

Still need to directly hook up the breadcrumbs to the questions object.