Debugged Firefox issues
Firefox was incorrectly displaying the page; in the end it boiled down
to innerText not being supported by the browsers and had to switch it
over to innerHTML to add text.

Also changed the elements over proper elements instead of custom
elements and adding an ID to them. However this did not correct the
issue.