Restructured question format and added functionality
Changed the questions format to be more human readable and added more
information to each question for more functionality.

Added ability to hide and show questions based on the choices of
previous questions.

Improved question tracking to keep-up with new features.
This includes keeping a count of questions available to answer and ones
answered; also removes them from this object if they are no longer
available to be answered.
Also updated the progress bar to accurately track the total available
questions and answered questions and visually show that.