Hooked up the summary estimate
Changed how the values of the answers are stored and recalled so that I
can easily find the dollar values associated to said answer.

Hooked up the summary to combine the different values for the selected
answers and display them once again.

Changed the Breadcrumb function, and separated the creation of the
breadcrumb from the changing of the active crumb.