Bug fixes, improvements, relation range
Adjusted the relation structure to be an object and need an array of
numbers to allow more than question answer to show the question.

Fleshed out the README.md more.

Bolded the question descriptions. moved the estimates into the progress
bar and fiddled with the font sizes.

Added many more notes.

There is a bug with questions 3 "layers" deep or more. if the third
layer is answered and then the original 1st layer question changed to
hide the 2nd layer; the 3rd layer questions remains but has it's value
wiped.