Adjusted progress bar logic
Changed the progress calculation to take all possible questions into
account rather than the ones able to be answered with the current
question answers.

Added an if statement after progress calculation that is similar to the
original progress calculation. Where if the count of questions answered
matches the number of question to answer, automatically update he
progress to 100% (because they wouldn't  have any more questions to
answer.