Added visual indicator for question
Added a red outline to the question dropdown when it is in need of
answering.

Also added a red outline to the calculate quote button when the progress
is 100% which goes away once pressed.

Added an outline to the progress bar itself when the final estimate is
shown.

These changes help to guide the users view through the application.