Disconnected the category system
Made the category system customizable so users are able to choose more
than services or disbursements or not have them at all.

The system creates separators based upon the types of the questions. and
is able to determine which spacer the question should be under and place
it at the bottom of the list for that spacer/category.

It is still a little buggy with more than two categories and if they are
not in order (questions are made into "test" and then go back to
services).