Added funeral price approximations
Added ability for users to setup an approximation value to be displayed
instead of the set amount.

Also added 2 approximation methods to choose from and the ability to add
more if they desire.

System defaults to showing the the true value if either the scope or
approxType value are not specified correctly.