Made adding fixed costs easier
Added an object to the formulas.js file called fixedCosts that assigns
costs to either services or disbursements.

The values added there are added to the final value at the end no matter
what questions are answered.

Also did something unnecessary in that made the property name a string;
this will make it easier later on down the line to display a summary of
the costs incurred.