Corrected issue #8
Re-organized the questions variable into an array instead of an object
so that I could refer to the questions position in the file rather than
relying on the order of it on the application.

This has caused me to re-code a fair few areas of the script.js code as
it now needs to refer to the question by it's position rather than a
unique ID.

There are still a few systems using the old reference method and will
need to be remedied. One in particular I found was the formula system.