Added scrolling to bottom and styling improvements
Generally improved applications look, focusing on mobile.

Corrected the bug where the background pattern would be redrawn when the
body ends (made the body = scrollHeight)

Added feature where the window scrolls to the bottom of the page when a
question is answered and a new one appears. This may cause annoyance if
the user has the application setup in such a way that the questions jump
around due to categories.

The approximation wasn't resetting each time it was calculated so made
the accounts default to 0 at the beginning of the GenSummary function.