Added color customization
Extrapolated most if not all the colors from the application into the
colors.js file, for issue #12.

I separated the file into 3 sections: constants, which allow the
repeated use of the color (variables); An object containing the
references the code uses to get colors; Custom styling code, allowing
the manipulation of individual aspects of the application (changing the
theme).

I was able to get most of the theme into the custom code area,  a few
things such as the highlighting effect is a little more complex and will
require the move to classes to allow it to be moved to the colors.js
file.