Implemented answer image and blurb feature
To fulfil issue #9.
Have implemented a system allowing the user to specify a blurb in the
option object of the questions that will be displayed when the question
is answered.

Have also implemented another system that works along side the blurb in
that it displays a image under the question when answered. The images
are stored in the images folder in the root directory. In order to point
the answer to the image enter the image property in much the same way as
the blurb, instead specifying the file name and extension to look for.

The two are not particularly 'pretty' at the moment and will need
adjusting later on. The code is also a bit 'get it done' so it could use
refactoring.