Added customization evaluation system
Created a system that loads in the custom code via a hidden HTML
objects. It then waits for the objects to load in their content and
decodes and parses the text though a evaluation function.

If the code is found to have errors (such as a missing comma), an alert
is shown describing the error and it's location.

Else the code is executed, ready to be used by the web application.

Also converted all the variables to window properties as it's easier to
use them on a global scope.

Also added more separation between the code by including more files
(fixedCosts.js, approximation.js)

This effectively solves issue #11.