Adjusted result calculation structure
Changed how the results of the summary is stored into objects. This is
so that it can be recalled at a later date should it be required to
construct a summary of sorts of how it came to the result figure.

I placed individual accounts (services, disbursements) into the accounts
object and in those individual accounts is a summary of each question
and it's answer and value (includes fixed costs and formula results.

I totaled each account's individual item into another object called
estimate where each account's total is stored.

I then totaled the individual accounts into the results variable and
then used that to come up the the end estimate value.