Worked out a bug with using locally stored version
Included the full window URL in the XHR request so that the files
location in the local storage is included so it loads correctly whether
started from a local file or a hosted website.

Also added a little more hints to the page to indicate that the
application is not functioning correctly.