Added email button
Added a button that redirects a mailto: , opening the browser's email
handler and fills in the subject and body with information about the
estimate; from there it is assumed the user can manipulate the
information to their liking.

In addition, I also added an extra property to each object in the
formulas object, this allows the summary function to add a text property
to the formula in what ever summary it generates, being more descriptive
than simply "catering" for example.