Updated Background coloring
Changed the CSS in a way so that the bodies background only displays
once and that the default HTML background is the same as the final
colour of the body background, giving a fluid appearance.

This corrects a bug that has been happening where the background color
repeats in certain circumstances.