Corrected Email Creation
Changed the way the URL is made for the mailto: email

made it join an dynamic array rather than simply appending static
strings to the URL to prevent the "&" being added before the subject,
despite there not being anything after the "?". This was causing the
subject to not be used when tested on Outlook mail client.